requirements.txt
python-telegram-bot[webhooks]==20.3
Flask
yt-dlp
openai==0.28
pytesseract
Pillow
pytube==15.0.0